# contrato-api/openapi.yaml
openapi: 3.1.0
info:
  title: ValidaLab API
  summary: Contrato de API único de ValidaLab — fuente de verdad entre frontend y backend.
  description: |
    Este documento OpenAPI es la **fuente de verdad única** del contrato de API de ValidaLab.
    Tanto `frontend/` como `backend/` desarrollan contra este documento sin conocer ni depender
    del código del otro paquete. Cualquier endpoint, payload o comportamiento de borde entre
    cliente y servidor se define **aquí primero** y solo después se implementa.

    Es un **único documento** (no fragmentado por módulos). La navegación por dominio se hace
    mediante los `tags` declarados más abajo (`usuarios`, `ideas`, `contactos`, `entrevistas`,
    `kpis`, `agente`, `proveedores`), que se corresponden con los módulos de dominio del backend.

    ## Convenciones transversales

    Estas reglas aplican a **todos** los endpoints. El resumen normativo vive también en el
    `CLAUDE.md` de la raíz; el detalle (catálogo de errores, esquemas concretos) vive aquí.

    ### Autenticación
    Todas las operaciones (salvo registro y login) requieren un **JWT** en
    `Authorization: Bearer <token>` (esquema `bearerAuth`). Una petición sin token válido
    devuelve `401` con el error `NO_AUTENTICADO`.

    ### Aislamiento multi-tenant por `owner_id` (RNF-04, RNF-05)
    ValidaLab es multi-tenant desde la primera versión. El `owner_id` se deriva **siempre** del
...

    ### Formato de error
...
    en este documento, no en `CLAUDE.md`.

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /usuarios/registro:
    post:
      tags: [usuarios]
      summary: Registrar una cuenta nueva
      description: Endpoint público. Crea la cuenta con rol `validador` y estado `activo` por defecto.
      operationId: registrarUsuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroUsuarioRequest'
      responses:
        '201':
          description: Cuenta creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '409':
          $ref: '#/components/responses/Conflicto'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
  /usuarios/login:
    post:
      tags: [usuarios]
      summary: Autenticarse y obtener la sesión
      description: |
        Endpoint público. Devuelve el `accessToken` en el cuerpo y entrega el `refreshToken` en
        una cookie `HttpOnly; Secure; SameSite=Strict` (ver `refreshCookie`). El cuerpo NO incluye
        el refresh token.
      operationId: iniciarSesion
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticación correcta.
          headers:
            Set-Cookie:
              description: |
                Cookie `refreshToken=<opaco>; HttpOnly; Secure; SameSite=Strict; Path=/usuarios;
                Max-Age=<TTL>`. El navegador la reenvía a `refresh` y `logout`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRespuesta'
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /usuarios/refresh:
    post:
      tags: [usuarios]
      summary: Renovar el token de acceso
      description: |
        Endpoint basado en la cookie de refresh (el `accessToken` puede haber expirado). Lee el
        `refreshToken` de la cookie `refreshToken`, lo **rota** (invalida el usado, emite uno
        nuevo) y devuelve un `TokenRespuesta` nuevo con una `Set-Cookie` rotada. No recibe cuerpo.
        Cookie ausente o refresh token inválido/expirado/revocado → `401`.
      operationId: refrescarToken
      security:
        - refreshCookie: []
      responses:
        '200':
          description: Token renovado.
          headers:
            Set-Cookie:
              description: Cookie `refreshToken` rotada (mismo formato que en login).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRespuesta'
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /usuarios/logout:
    post:
      tags: [usuarios]
      summary: Cerrar sesión
      description: |
        Endpoint basado en la cookie de refresh; no exige `accessToken` válido (así el cierre
        funciona aunque el access token haya expirado). Lee el `refreshToken` de la cookie, lo
        invalida y **limpia la cookie** en la respuesta. No recibe cuerpo. Es idempotente: sin
        cookie o con una ya revocada responde `204` igualmente.
      operationId: cerrarSesion
      security:
        - refreshCookie: []
      responses:
        '204':
          description: Sesión cerrada; el `refreshToken` deja de ser válido y la cookie se limpia.
          headers:
            Set-Cookie:
              description: Cookie `refreshToken` limpiada (expiración inmediata).
              schema:
                type: string
  /usuarios/yo:
    get:
      tags: [usuarios]
      summary: Consultar el perfil propio
      description: Resuelve el usuario a partir del token; no recibe `id`.
      operationId: obtenerPerfilPropio
      responses:
        '200':
          description: Perfil del usuario autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NoAutenticado'
    patch:
      tags: [usuarios]
      summary: Actualizar el perfil propio
      description: Solo permite cambiar datos del propio perfil (p. ej. `nombre`); no el `rol` ni el `estado`.
      operationId: actualizarPerfilPropio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarPerfilRequest'
      responses:
        '200':
          description: Perfil actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
  /usuarios:
    get:
      tags: [usuarios]
      summary: Listar cuentas (solo administrador)
      description: Requiere rol `administrador` (RBAC). Un `validador` recibe `403 ACCESO_DENEGADO`.
      operationId: listarUsuarios
      parameters:
        - $ref: '#/components/parameters/pagina'
        - $ref: '#/components/parameters/porPagina'
      responses:
        '200':
          description: Página de cuentas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuariosPaginados'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
  /usuarios/{id}:
    get:
      tags: [usuarios]
      summary: Consultar una cuenta (solo administrador)
      description: Requiere rol `administrador` (RBAC). Un `validador` recibe `403 ACCESO_DENEGADO`.
      operationId: obtenerUsuario
      parameters:
        - $ref: '#/components/parameters/idUsuario'
      responses:
        '200':
          description: Cuenta solicitada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
  /usuarios/{id}/rol:
    patch:
      tags: [usuarios]
      summary: Cambiar el rol de una cuenta (solo administrador)
      description: Requiere rol `administrador` (RBAC). Un `validador` recibe `403 ACCESO_DENEGADO`.
      operationId: cambiarRolUsuario
      parameters:
        - $ref: '#/components/parameters/idUsuario'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambiarRolRequest'
      responses:
        '200':
          description: Rol actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
  /usuarios/{id}/estado:
    patch:
      tags: [usuarios]
      summary: Cambiar el estado de una cuenta (solo administrador)
      description: Requiere rol `administrador` (RBAC). Permite suspender/activar una cuenta. Un `validador` recibe `403 ACCESO_DENEGADO`.
      operationId: cambiarEstadoUsuario
      parameters:
        - $ref: '#/components/parameters/idUsuario'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambiarEstadoRequest'
      responses:
        '200':
          description: Estado actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT emitido por ValidaLab tras el login. El `owner_id` del usuario se deriva del token;
        nunca se acepta como entrada del cliente.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: |
        Refresh token opaco transportado en una cookie `HttpOnly; Secure; SameSite=Strict` con
        `Path` acotado a las rutas de sesión. El navegador la envía automáticamente a
        `/usuarios/refresh` y `/usuarios/logout`; el JavaScript del cliente NUNCA la lee ni la
        escribe. La emite `login`, la rota `refresh` y la limpia `logout`.
  parameters:
    pagina:
      name: pagina
      in: query
      required: false
      description: Número de página (base 1).
      schema:
        type: integer
        minimum: 1
        default: 1
    porPagina:
      name: porPagina
      in: query
      required: false
      description: Cantidad de elementos por página.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    idUsuario:
      name: id
      in: path
      required: true
      description: Identificador de la cuenta de usuario.
      schema:
        type: string
        format: uuid
...
  schemas:
    CodigoError:
      type: string
      description: |
        Catálogo estable de códigos de error de ValidaLab. El cliente puede ramificar su lógica
        sobre estos valores; el texto de `mensaje` es informativo y puede cambiar.
      enum:
        - VALIDACION_FALLIDA        # 400 — payload inválido o reglas de negocio no satisfechas
        - NO_AUTENTICADO            # 401 — falta token o es inválido/expirado
        - ACCESO_DENEGADO           # 403 — el recurso pertenece a otro owner_id (aislamiento)
        - RECURSO_NO_ENCONTRADO     # 404 — el recurso no existe o no es visible para el usuario
        - CONFLICTO                 # 409 — conflicto de estado (p. ej. transición de outreach inválida)
        - ENTREVISTA_SIN_VINCULO    # 422 — entrevista sin idea o contacto válidos del mismo usuario (RNF-14)
        - API_KEY_INVALIDA          # 422 — la API key BYOK no validó contra el proveedor (RF-20)
        - SALIDA_AGENTE_INVALIDA    # 502 — la salida del agente no cumplió su esquema Zod tras reintentos (RF-AG-03)
        - PROVEEDOR_IA_NO_DISPONIBLE # 503 — el proveedor de IA elegido no respondió (RNF-08)
        - LIMITE_TASA               # 429 — demasiadas solicitudes
        - ERROR_INTERNO             # 500 — error inesperado del servidor
    Error:
      type: object
      description: Sobre estándar de error de toda la API.
      required: [codigo, mensaje]
      properties:
        codigo:
          $ref: '#/components/schemas/CodigoError'
        mensaje:
          type: string
          description: Descripción legible del error (informativa, no estable).
        detalles:
          type: array
          description: Errores de validación campo a campo, cuando aplica.
          items:
            type: object
            properties:
              campo:
                type: string
              problema:
                type: string
    Paginacion:
      type: object
      description: Metadatos de paginación que acompañan a toda colección.
      required: [pagina, porPagina, total, totalPaginas]
      properties:
        pagina:
          type: integer
          description: Página actual (base 1).
        porPagina:
          type: integer
          description: Elementos por página solicitados.
        total:
          type: integer
          description: Total de elementos disponibles para el usuario autenticado.
        totalPaginas:
          type: integer
          description: Total de páginas según `total` y `porPagina`.
    RespuestaPaginada:
      type: object
      description: |
        Sobre de respuesta para colecciones. Los endpoints concretos reemplazan/estrechan
        `datos` con el tipo de recurso correspondiente.
      required: [datos, paginacion]
      properties:
        datos:
          type: array
          items: {}
        paginacion:
          $ref: '#/components/schemas/Paginacion'

    # --- Módulo usuarios (épica E0) ---
    Rol:
      type: string
      description: Rol del usuario para RBAC.
      enum:
        - validador      # usuario primario; gestiona su propio portafolio aislado
        - administrador  # gestiona cuentas, roles y estado de otros usuarios
    EstadoUsuario:
      type: string
      description: Estado de la cuenta. Una cuenta `suspendido` no puede autenticarse.
      enum:
        - activo
        - suspendido
    Usuario:
      type: object
      description: |
        Recurso de cuenta de usuario. NUNCA incluye la contraseña ni su hash en ninguna respuesta.
      required: [id, email, nombre, rol, estado, fechaCreacion]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nombre:
          type: string
        rol:
          $ref: '#/components/schemas/Rol'
        estado:
          $ref: '#/components/schemas/EstadoUsuario'
        fechaCreacion:
          type: string
          format: date-time
    RegistroUsuarioRequest:
      type: object
      required: [email, nombre, password]
      properties:
        email:
          type: string
          format: email
        nombre:
          type: string
          minLength: 1
        password:
          type: string
          format: password
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    TokenRespuesta:
      type: object
      description: |
        Sesión emitida en login y refresh. El `refreshToken` NO viaja aquí: se entrega en la
        cookie `HttpOnly` `refreshToken` (ver el esquema de seguridad `refreshCookie`). El cuerpo
        solo lleva el `accessToken`, que el cliente usa en `Authorization: Bearer` y mantiene en
        memoria.
      required: [accessToken, tokenTipo, expiraEn, usuario]
      properties:
        accessToken:
          type: string
          description: "JWT de acceso para `Authorization: Bearer`."
        tokenTipo:
          type: string
          enum: [Bearer]
          default: Bearer
        expiraEn:
          type: integer
          description: Segundos de vigencia del `accessToken`.
        usuario:
          $ref: '#/components/schemas/Usuario'
    ActualizarPerfilRequest:
      type: object
      description: Cambios permitidos sobre el perfil propio (no incluye rol ni estado).
      properties:
        nombre:
          type: string
          minLength: 1
    CambiarRolRequest:
      type: object
      required: [rol]
      properties:
        rol:
          $ref: '#/components/schemas/Rol'
    CambiarEstadoRequest:
      type: object
      required: [estado]
      properties:
        estado:
          $ref: '#/components/schemas/EstadoUsuario'
    UsuariosPaginados:
      description: Página de cuentas de usuario (sobre `RespuestaPaginada` con `datos` de `Usuario`).
      allOf:
        - $ref: '#/components/schemas/RespuestaPaginada'
        - type: object
          properties:
            datos:
              type: array
              items:
                $ref: '#/components/schemas/Usuario'

    # --- Módulo ideas (épica E1) ---
...
  responses:
    NoAutenticado:
      description: Falta el token o es inválido/expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccesoDenegado:
      description: El recurso pertenece a otro `owner_id` (regla de aislamiento multi-tenant).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoEncontrado:
      description: El recurso no existe o no es visible para el usuario autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidacionFallida:
      description: El payload no cumple las reglas de validación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflicto:
      description: Conflicto de estado (p. ej. un `email` ya registrado).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...